requirements.txt
requests==2.32.3
pytest==8.3.5
pytest-html==4.1.1
pytest-cov==6.1.0
jsonschema==4.23.0
pydantic==2.10.6
python-dotenv==1.0.1
